==> 2. Text Extraction.ipynb <==
import pandas as pd
import PyPDF2
import os
import re

def text_extraction(file_path):
    pdf_text = ""
    with open(file_path, 'rb') as pdf:
        reader = PyPDF2.PdfFileReader(pdf, strict=False)
        
        for page in reader.pages:
            content = page.extract_text()
            pdf_text += " " + content
    
    return pdf_text

root_path = "data\ML Profiles"
pdfs = os.listdir(root_path)
texts = []
label = []
for profile in pdfs:
    file_path = os.path.join(root_path, profile)
    profile_text = text_extraction(file_path)
    texts.append(profile_text)
    label.append(1)

root_path = "data\Painting Profiles"
pdfs_aux = os.listdir(root_path)
pdfs += pdfs_aux
for profile in pdfs_aux:
    file_path = os.path.join(root_path, profile)
    profile_text = text_extraction(file_path)
    texts.append(profile_text)
    label.append(0)

profiles_df = pd.DataFrame()
profiles_df["PDF"] = pdfs
profiles_df["Text"] = texts
profiles_df["label"] = label
profiles_df.to_csv("data/Profiles.csv", index=False)

def extract_summary(profiles_df_text):
    text = profiles_df_text.replace("\n",". ")
    start = ". Summary."
    end = ". Experience."

    start_index = text.find(start)
    end_index = text.find(end)

    summary = text[start_index + len(start):end_index]
    
    return summary

def extract_experience(profiles_df_text):
    text = profiles_df_text.replace("\n",". ")
    start = ". Experience."
    end = ". Education."

    start_index = text.find(start)
    end_index = text.find(end)

    experience = text[start_index + len(start):end_index]
    
    return experience

def extract_education(profiles_df_text):
    text = profiles_df_text.replace("\n",". ")
    start = ". Education."

    start_index = text.find(start)

    experience = text[start_index + len(start):]
    
    return experience

profiles_df["Summary"] = [extract_summary(x) for x in profiles_df["Text"]]
profiles_df["Experience"] = [extract_experience(x) for x in profiles_df["Text"]]
profiles_df["Education"] = [extract_education(x) for x in profiles_df["Text"]]
profiles_df.to_csv("data/Profiles.csv", index=False)

profiles_df = pd.read_csv("data/Profiles.csv")

profiles_df.head()
