==> 3. Train Test Split.ipynb <==
from sklearn.model_selection import train_test_split
import pandas as pd

profiles_df = pd.read_csv("data/Profiles.csv")
profiles_df.head()

columns = profiles_df.columns.to_list()
columns.remove('label')

X = profiles_df[columns]
y = profiles_df['label']

X_train, X_test, y_train, y_test = train_test_split(X, y, test_size=0.2, random_state=42)

X_train.to_csv("data/X_train.csv", index=False)
y_train.to_csv("data/y_train.csv", index=False)
X_test.to_csv("data/X_test.csv", index=False)
y_test.to_csv("data/y_test.csv", index=False)
