==> 4d. BagOfWords Vectorization.ipynb <==
import numpy as np
import nltk
from nltk.corpus import stopwords
from nltk.tokenize import word_tokenize
from sklearn.feature_extraction.text import HashingVectorizer
import pandas as pd
import pickle
from scipy import sparse
import os

os.chdir("C:\\Users\wwwdz\OneDrive\Desktop\Spring Semester 2023\AI\FinalProject\LinkedInMatch")
os.listdir()

train_profiles = pd.read_csv("data/X_train.csv")
test_profiles = pd.read_csv("data/X_test.csv")

nltk.download('punkt')
nltk.download('stopwords')
stop_words = set(stopwords.words('english'))

def remove_stopwords(text):
    tokens = word_tokenize(text.lower())
    tokens_wo_stopwords = [t for t in tokens if t not in stop_words]
    text_no_stopwords = " ".join(tokens_wo_stopwords)
    return text_no_stopwords

train_texts = train_profiles["Text"]
train_texts_no_stopwords = []
for text in train_texts:
    train_texts_no_stopwords.append(remove_stopwords(text))

test_texts = test_profiles["Text"]
test_texts_no_stopwords = []
for text in test_texts:
    test_texts_no_stopwords.append(remove_stopwords(text))

corpus = train_texts_no_stopwords
# Vectorize the corpus using feature hashing and BoW
vectorizer = HashingVectorizer()
X = vectorizer.fit_transform(corpus)
train_bow_feature_vectors = vectorizer.fit_transform(train_texts_no_stopwords)
test_bow_feature_vectors = vectorizer.fit_transform(test_texts_no_stopwords)

sparse.save_npz("data/train_bow_feature_vectors.npz", train_bow_feature_vectors)
sparse.save_npz("data/test_bow_feature_vectors.npz", test_bow_feature_vectors)

pickle.dump(X, open("data/bow.pickle", "wb"))
pickle.dump(train_bow_feature_vectors, open("data/train_bow_feature_vectors.pickle", "wb"))
pickle.dump(test_bow_feature_vectors, open("data/test_bow_feature_vectors.pickle", "wb"))
