==> 5b. BERT Logistic Regression.ipynb <==
from sklearn.linear_model import LogisticRegression
import pandas as pd
import torch

X_train = torch.load("data/train_summaries.pt")
y_train = pd.read_csv("data/y_train.csv")
X_test = torch.load("data/test_summaries.pt")
y_test = pd.read_csv("data/y_test.csv")
print(len(X_train), len(y_train))

train_profiles_features = []
for f in X_train:
    vector = f[0]
    train_profiles_features.append(vector)

test_profiles_features = []
for f in X_test:
    vector = f[0]
    test_profiles_features.append(vector)

clf = LogisticRegression(random_state=0, max_iter=4000).fit(train_profiles_features, y_train['label'])

clf.score(train_profiles_features,y_train['label'])

clf.score(test_profiles_features,y_test['label'])
