==> 5d. BagOfWords Logistic Regression.ipynb <==
from sklearn.linear_model import LogisticRegression
import pandas as pd
from scipy import sparse
import os
from sklearn.linear_model import SGDClassifier

os.chdir("C:\\Users\wwwdz\OneDrive\Desktop\Spring Semester 2023\AI\FinalProject\LinkedInMatch")
os.listdir()

X_train = sparse.load_npz("data/train_bow_feature_vectors.npz")
X_test = sparse.load_npz("data/test_bow_feature_vectors.npz")

X_train.shape

X_test

y_train = pd.read_csv("data/y_train.csv")
y_test = pd.read_csv("data/y_test.csv")

y_test[:20]

model = LogisticRegression(random_state=0, ).fit(X_train, y_train['label'])
model.score(X_test,y_test['label'])

sgdmodel = SGDClassifier(loss='log', penalty='l2', alpha=0.0001, max_iter=100, tol=1e-3, learning_rate='adaptive', eta0=0.1)
sgdmodel.fit(X_train, y_train['label'])
model.score(X_test,y_test['label'])
